Task 3: Fix inode/device truncation and re-enable gosec G115
dcfh stored Dev/Ino as uint32 and keyed hardlink dedup on a truncated
[2]uint32, silently collapsing distinct files as hardlinks on large-inode
filesystems (XFS, Btrfs, large ext4). This widens Dev/Ino to 64-bit and, to
make the change single-sited, first encapsulates the entire versioned on-disk
format behind one module (pkg/format), then re-enables the gosec G115 gate
Task 2 had deferred.

Delivered across three subtasks:
- 3.1 (chore): extract pkg/format as the single, cycle-free owner of the
  on-disk layout; migrate core + dcfhfix onto it; delete both dcfhfix
  duplicates and the parallel offset table. No width/version/behaviour change.
- 3.2 (feature): add the owned version->DecodeStrategy dispatch seam and give
  pkg/format ownership of the write version. No format change.
- 3.3 (bugfix): widen DevID/Inode to uint64 (on-disk format v4); legacy v2/v3
  read via a bounds-checked heap transcode; dupes key [2]uint64; ingest
  truncation removed; gosec G115 re-enabled and whole-tree clean.

All five success criteria verified end-to-end on the merged branch: full-width
dev/ino + full-width dupes key, single-module format ownership, v2/v3->v4
heap-transcode round-trip, G115 active with 0 whole-tree findings, full race
suite green. Five pre-existing latent defects found and fixed in-flight via
planned gates; 0 escaped.

Closes the Very High inode/device-truncation backlog item.

+++ pkg/binary_entry_index_file_mmap.go
@@ -105,7 +105,7 @@ func (ime *BEIndexFileMmapEntry) MTimeWall() (uint64, error) {
 }
 
 // Dev returns the device ID
-func (ime *BEIndexFileMmapEntry) Dev() (uint32, error) {
+func (ime *BEIndexFileMmapEntry) Dev() (uint64, error) {
 	ime.RLock()
 	defer ime.RUnlock()
 
@@ -118,7 +118,7 @@ func (ime *BEIndexFileMmapEntry) Dev() (uint32, error) {
 }
 
 // Ino returns the inode number
-func (ime *BEIndexFileMmapEntry) Ino() (uint32, error) {
+func (ime *BEIndexFileMmapEntry) Ino() (uint64, error) {
 	ime.RLock()
 	defer ime.RUnlock()
